Big stuff
Cosmic is a mostly written from the ground up frontend that I have been working on to replace the current modified deviantart 2006 frontend. 

As I have not been properly tracking files this is just a dump of what is already complete.

The frontend is quite near completion though still unfinished.