Show message on post failure and direct user to post
When a thread or comment is empty a message will appear notifying you about this issue rather than allowing a blank post to be created.

Users can now quickly jump to messages using /p/<post_id> (goto_post) and said post will be highlighted until the page is reloaded. The page will also try to jump to this message with use of "#" <a> links, this uses the message ID to jump 

IE: "/b/0/0#5" will take us to board 0, thread 0, comment with the ID of 5.